Share structured docstring parsing components
