+++ crates/ty_ide/src/docstring/document/google.rs
@@ -35,12 +35,12 @@ use ruff_python_stdlib::identifiers::is_identifier;
 use ruff_python_trivia::Cursor;
 use ruff_text_size::{TextRange, TextSize};
 
-use super::SectionKind;
 use super::preformatted::PreformattedBlockScanner;
 use super::syntax::{
-    ParsedLine, consume_quoted_string, container_block_end, indentation, parsed_lines,
-    split_once_at_top_level_colon, split_trailing_parenthetical, starts_with_markdown_list_item,
+    ParsedLine, consume_quoted_string, container_block_end, indentation, is_dotted_identifier,
+    parsed_lines, split_once_at_top_level_colon, split_trailing_parenthetical,
 };
+use super::{DescriptionBuilder, HeaderKind, SectionKind};
 
 /// Returns parameter documentation from recognized Google-style parameter sections.
 ///
@@ -73,95 +73,15 @@ pub(in crate::docstring) fn sections(source: &str) -> impl Iterator<Item = Secti
 }
 
 /// A recognized Google-style docstring section.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(in crate::docstring) struct Section {
-    kind: SectionKind,
-    range: TextRange,
-    body: SectionBody,
-}
+pub(in crate::docstring) type Section = super::Section<String>;
 
-impl Section {
-    /// Returns the section kind.
-    pub(in crate::docstring) const fn kind(&self) -> SectionKind {
-        self.kind
-    }
-
-    /// Returns the section's source range.
-    pub(in crate::docstring) const fn range(&self) -> TextRange {
-        self.range
-    }
-
-    /// Consumes this section and returns its fragments when it can be rendered structurally.
-    pub(in crate::docstring) fn into_renderable_fragments(self) -> Option<Vec<BodyFragment>> {
-        let SectionBody::Parsed {
-            fragments,
-            has_structural_ambiguity,
-        } = self.body
-        else {
-            return None;
-        };
-        (!has_structural_ambiguity).then_some(fragments)
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-enum SectionBody {
-    /// A body parsed into semantic fragments.
-    Parsed {
-        fragments: Vec<BodyFragment>,
-        /// Whether the body's structure is ambiguous.
-        has_structural_ambiguity: bool,
-    },
-    /// A body whose contents were not parsed.
-    Opaque,
-}
-
-impl SectionBody {
-    /// Creates an unambiguous body containing the description as a single prose fragment.
-    fn from_prose(description: String) -> Self {
-        let fragments = (!description.is_empty())
-            .then_some(BodyFragment::Prose(description))
-            .into_iter()
-            .collect();
-        Self::Parsed {
-            fragments,
-            has_structural_ambiguity: false,
-        }
-    }
-
-    fn into_fragments(self) -> Vec<BodyFragment> {
-        match self {
-            Self::Parsed { fragments, .. } => fragments,
-            Self::Opaque => Vec::new(),
-        }
-    }
-}
+type SectionBody = super::SectionBody<String>;
 
 /// One parsed fragment in a Google section body.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(in crate::docstring) enum BodyFragment {
-    /// Section-level prose that is not attached to a named item.
-    Prose(String),
-    /// A named item and its description.
-    Item(Item),
-}
+pub(in crate::docstring) type BodyFragment = super::BodyFragment<String>;
 
 /// A named item in a Google section.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(in crate::docstring) struct Item {
-    display_name: String,
-    ty: Option<String>,
-    description: String,
-}
-
-impl Item {
-    /// Consumes this item and returns its display parts.
-    pub(in crate::docstring) fn into_display_name_type_and_description(
-        self,
-    ) -> (String, Option<String>, String) {
-        (self.display_name, self.ty, self.description)
-    }
-}
+pub(in crate::docstring) type Item = super::Item<String>;
 
 /// Splits a display name from a trailing parenthesized type.
 ///
@@ -552,11 +472,6 @@ impl<'a> SectionBuilder<'a> {
     }
 }
 
-/// Returns whether every component of `name` is a Python identifier.
-fn is_dotted_identifier(name: &str) -> bool {
-    !name.is_empty() && name.split('.').all(is_identifier)
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 struct Header {
     kind: HeaderKind,
@@ -569,12 +484,6 @@ struct Header {
 enum HeaderForm {
     Section,
     Inline,
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum HeaderKind {
-    Structured(SectionKind),
-    Opaque,
 }
 
 impl HeaderKind {
@@ -754,74 +663,6 @@ impl<'a> ItemBuilder<'a> {
             ty: self.ty.map(str::to_string),
             description: self.description.finish(),
         }
-    }
-}
-
-#[derive(Default)]
-struct DescriptionBuilder<'a> {
-    inline: Option<&'a str>,
-    continuation_lines: Vec<&'a str>,
-}
-
-impl<'a> DescriptionBuilder<'a> {
-    fn with_inline(inline: &'a str) -> Self {
-        let inline = inline.trim();
-        Self {
-            inline: (!inline.is_empty()).then_some(inline),
-            continuation_lines: Vec::new(),
-        }
-    }
-
-    fn push_line(&mut self, line: &'a str) {
-        // Keep a leading list item with the block so that its indentation establishes the baseline
-        // for nested items. Ordinary first lines use the allocation-free inline representation.
-        if self.inline.is_none()
-            && self.continuation_lines.is_empty()
-            && !starts_with_markdown_list_item(line.trim_start())
-        {
-            self.inline = Some(line.trim());
-        } else {
-            self.push_continuation(line);
-        }
-    }
-
-    fn push_continuation(&mut self, line: &'a str) {
-        self.continuation_lines.push(line);
-    }
-
-    fn finish(mut self) -> String {
-        if self.continuation_lines.is_empty() {
-            return self.inline.map_or_else(String::new, str::to_string);
-        }
-
-        let continuation_indent = self
-            .continuation_lines
-            .iter()
-            .filter(|line| !line.trim().is_empty())
-            .map(|line| indentation(line))
-            .min()
-            .unwrap_or_default();
-        for line in &mut self.continuation_lines {
-            *line = if line.trim().is_empty() {
-                ""
-            } else {
-                strip_indentation(line, continuation_indent).trim_end()
-            };
-        }
-
-        if let Some(inline) = self.inline {
-            self.continuation_lines.insert(0, inline);
-        }
-        let lines = self.continuation_lines;
-
-        let Some(start) = lines.iter().position(|line| !line.is_empty()) else {
-            return String::new();
-        };
-        let end = lines
-            .iter()
-            .rposition(|line| !line.is_empty())
-            .map_or(start, |index| index + 1);
-        lines[start..end].join("\n")
     }
 }
 
@@ -1049,41 +890,6 @@ fn is_attribute_display_name(display_name: &str) -> bool {
     display_name
         .split(',')
         .all(|name| is_dotted_identifier(name.trim()))
-}
-
-fn strip_indentation(line: &str, width: TextSize) -> &str {
-    let mut indentation_width = TextSize::default();
-    for (index, char) in line.char_indices() {
-        let next_indentation_width = match char {
-            ' ' => indentation_width + TextSize::new(1),
-            '\t' => TextSize::new((indentation_width.to_u32() / 8 + 1) * 8),
-            _ => return &line[index..],
-        };
-
-        if next_indentation_width > width {
-            return &line[index..];
-        }
-
-        indentation_width = next_indentation_width;
-        if indentation_width == width {
-            return &line[index + char.len_utf8()..];
-        }
-    }
-
-    ""
-}
-
-impl HeaderKind {
-    fn is_parameter_section(self) -> bool {
-        matches!(
-            self,
-            Self::Structured(
-                SectionKind::Parameters
-                    | SectionKind::KeywordArguments
-                    | SectionKind::OtherParameters
-            )
-        )
-    }
 }
 
 #[cfg(test)]
